split tests file
